+++ app/src/main/java/ru/netology/myapp/dao/PostRemoteKeyDao.kt
@@ -12,10 +12,10 @@ interface PostRemoteKeyDao {
     suspend fun isEmpty(): Boolean
 
     @Query("SELECT max(key) FROM PostRemoteKeyEntity")
-    suspend fun max(): Long?
+    suspend fun max(): Int?
 
     @Query("SELECT min(key) FROM PostRemoteKeyEntity")
-    suspend fun min(): Long?
+    suspend fun min(): Int?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(key: PostRemoteKeyEntity)
